backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.8
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
cloudinary==1.40.0
Pillow==10.3.0
slowapi==0.1.9
python-dotenv==1.0.1
httpx==0.27.0
email-validator==2.1.1
